change operator types to double to support divs

+++ ChainOfResponsibility/src/main/java/com/miguelvela/MathOperation.java
@@ -1,21 +1,21 @@
 package com.miguelvela;
 
 public class MathOperation {
-    private int operator1;
-    private int operator2;
+    private double operator1;
+    private double operator2;
     private String operation;
 
-    public MathOperation(int number1, int number2, String operation) {
+    public MathOperation(double number1, double number2, String operation) {
         this.operator1 = number1;
         this.operator2 = number2;
         this.operation = operation;
     }
 
-    public int getOperator1() {
+    public double getOperator1() {
         return operator1;
     }
 
-    public int getOperator2() {
+    public double getOperator2() {
         return operator2;
     }
